requirements.txt
Flask==2.0.3
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.0.3
mysql-connector-python==9.1.0
flask_cors==5.0.0
gunicorn==19.9.0
